requirements.txt
torch
torchsummary
torchvision
numpy
Pillow
PyYAML
requests
tqdm
opencv-python
pycocotools